redo config system to allow for bin targets
