+++ platforms/allwinner-d1/config/lichee-rv.toml
