+++ platforms/melpomene/Cargo.toml
@@ -18,6 +18,7 @@ melpo-config = { path = "./melpo-config" }
 serde = { version = "1.0.178", features = ["derive"] }
 mnemos-config = { path = "../../source/config", features = ["use-std"] }
 mnemos-kernel = { package = "mnemos", path = "../../source/kernel" }
+miette = "5.10.0"
 
 [dependencies]
 serde = { version = "1.0.178", features = ["derive"] }
