+++ platforms/melpomene/build.rs
@@ -1,3 +1,3 @@
 fn main() {
-    mnemos_config::buildtime::render_project::<melpo_config::PlatformConfig>("melpo.toml").unwrap();
+    mnemos_config::buildtime::render_file::<melpo_config::PlatformConfig>("melpo.toml").unwrap();
 }
