+++ source/config/src/lib.rs
@@ -58,10 +58,14 @@ pub struct MnemosConfig<Platform> {
     pub platform: Platform,
 }
 
+pub const CONFIG_DIR_VAR: &str = "MNEMOS_CONFIG_DIR";
+pub const CONFIG_FILE_VAR: &str = "MNEMOS_CONFIG";
+
 /// Tools intended for use in build.rs scripts
 #[cfg(feature = "use-std")]
 pub mod buildtime {
-    use std::{io::Write, path::PathBuf};
+    const OUT_DIR: &str = "OUT_DIR";
+    use std::{fs, io::Write, path::Path};
 
     use super::*;
     use miette::{Context, IntoDiagnostic, Result};
@@ -81,28 +85,116 @@ pub mod buildtime {
         postcard::to_stdvec(&mc).into_diagnostic()
     }
 
-    /// Load a configuration file from the given path, will be made available
-    /// to the main platform binary when they call [load_configuration!()].
-    pub fn render_project<Platform>(path: &str) -> Result<()>
+    /// Render all configuration files in the given directory.
+    pub fn render_all<Platform>(path: impl AsRef<Path>) -> Result<()>
     where
         Platform: Serialize + DeserializeOwned + 'static,
     {
-        let cfg = std::fs::read_to_string(path)
+        let path = path.as_ref();
+        let path_disp = path.display();
+        let out_dir = std::env::var(OUT_DIR)
             .into_diagnostic()
-            .wrap_err_with(|| format!("Failed to find input config file '{path}'"))?;
-        let c: MnemosConfig<Platform> = from_toml(&cfg)?;
+            .wrap_err("Failed to read '{OUT_DIR}' env variable")?;
 
-        let out_dir = std::env::var("OUT_DIR").into_diagnostic()?;
-        let mut out = PathBuf::from(out_dir);
+        println!("cargo:rerun-if-changed={path_disp}");
+        println!("cargo:rustc-env={CONFIG_DIR_VAR}={out_dir}");
 
-        out.push("mnemos-config.postcard");
-        let bin_cfg = to_postcard(&c)?;
-        let mut f = std::fs::File::create(&out).unwrap();
-        f.write_all(&bin_cfg).unwrap();
-        println!("cargo:rustc-env=MNEMOS_CONFIG={}", out.display());
-        println!("cargo:rerun-if-changed={path}");
+        println!("[mnemos_config] {OUT_DIR}={out_dir}");
+
+        (|| {
+            let mut rendered_any = false;
+            println!("[mnemos_config] rendering configs in '{path_disp}'...");
+            for entry in fs::read_dir(path)
+                .into_diagnostic()
+                .wrap_err_with(|| format!("Failed to read config file directory '{path_disp}'"))?
+            {
+                let entry = match entry {
+                    Ok(e) => e,
+                    Err(e) => {
+                        println!("cargo:warning=Error reading config dir entry: {e}");
+                        continue;
+                    }
+                };
+
+                let entry_path = entry.path();
+                let epath_disp = entry_path.display();
+                println!("[mnemos_config] {epath_disp}?");
+
+                if entry
+                    .metadata()
+                    .into_diagnostic()
+                    .wrap_err_with(|| format!("Failed to read metadata for '{epath_disp}'"))?
+                    .is_dir()
+                {
+                    println!("[mnemos_config] {epath_disp} -> not a file; skipping");
+                    continue;
+                }
+
+                if entry_path
+                    .extension()
+                    .map(|e| e == ".toml")
+                    .unwrap_or(false)
+                {
+                    println!("[mnemos_config] {epath_disp} -> not TOML; skipping");
+                    continue;
+                }
+
+                render_file_to::<Platform>(&entry_path, &out_dir)?;
+                rendered_any = true;
+            }
+
+            miette::ensure!(
+                rendered_any,
+                "No config files (paths ending in '.toml') found in directory!"
+            );
+
+            Ok(())
+        })()
+        .wrap_err_with(|| format!("Failed to render config directory '{path_disp}'"))?;
 
         Ok(())
+    }
+
+    /// Load a configuration file from the given path, will be made available
+    /// to the main platform binary when they call [include_config!()].
+    pub fn render_file<Platform>(path: impl AsRef<Path>) -> Result<()>
+    where
+        Platform: Serialize + DeserializeOwned + 'static,
+    {
+        let out_dir = std::env::var(OUT_DIR)
+            .into_diagnostic()
+            .wrap_err("Failed to read '{OUT_DIR}' env variable")?;
+        println!("[mnemos_config] {OUT_DIR}={out_dir}");
+        render_file_to::<Platform>(path, out_dir)
+    }
+
+    fn render_file_to<Platform>(path: impl AsRef<Path>, out: impl AsRef<Path>) -> Result<()>
+    where
+        Platform: Serialize + DeserializeOwned + 'static,
+    {
+        let path = path.as_ref();
+        let path_disp = path.display();
+        println!("[mnemos_config] rendering config file '{path_disp}'...");
+
+        (|| {
+            let filename = path
+                .file_name()
+                .ok_or_else(|| miette::miette!("Path has no filename!"))?;
+
+            let cfg = std::fs::read_to_string(path).into_diagnostic()?;
+            let c: MnemosConfig<Platform> = from_toml(&cfg)?;
+
+            let mut out = out.as_ref().join(filename);
+            out.set_extension("postcard");
+            let bin_cfg = to_postcard(&c)?;
+            let mut f = std::fs::File::create(&out).into_diagnostic()?;
+            f.write_all(&bin_cfg).into_diagnostic()?;
+            println!("cargo:rustc-env={CONFIG_FILE_VAR}={}", out.display());
+            println!("cargo:rerun-if-changed={path_disp}");
+
+            Ok::<_, miette::Report>(())
+        })()
+        .wrap_err_with(|| format!("Failed to render config file '{path_disp}'"))
     }
 }
 
@@ -128,7 +220,12 @@ pub mod runtime {
 ///
 /// Should be called with the type of your platform specific type
 #[macro_export]
-macro_rules! load_configuration {
+macro_rules! include_config {
+    ($platform: ty, $name: literal) => {{
+        const MNEMOS_CONFIG: &[u8] =
+            include_bytes!(concat!(env!("MNEMOS_CONFIG_DIR"), "/", $name, ".postcard"));
+        $crate::runtime::from_postcard::<$platform>(MNEMOS_CONFIG)
+    }};
     ($platform: ty) => {{
         const MNEMOS_CONFIG: &[u8] = include_bytes!(env!("MNEMOS_CONFIG"));
         $crate::runtime::from_postcard::<$platform>(MNEMOS_CONFIG)
